requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
jupyter>=1.0.0 
